remove utils/redirect module
